third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/animation-duration-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'animation-duration' to a time: -3.14ms
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
Harness: the test ran to completion.
